Data.Time

Description

Quick Start

Use these types for time regardless of location (not caring about leap-seconds):

UTCTime for actual times

NominalDiffTime for differences between times, i.e. durations

Use these types for the ways people refer to time and time differences:

Day for something like June 27th 2017

DayOfWeek for something like Tuesday

Month for something like August 2021

QuarterOfYear for something like Q2

Quarter for something like Q2 of 2023

TimeOfDay for something like 5pm

LocalTime for a Day with a TimeOfDay

TimeZone for a time zone offset (not actually the time zone itself) like -0700

ZonedTime for a LocalTime with a TimeZone

CalendarDiffDays for something like 6 years, 1 month and 5 days

CalendarDiffTime for something like 6 years, 1 month, 5 days, 3 hours, 7 minutes and 25.784 seconds

Use this for low-latency timing:

SystemTime

These are less commonly needed:

AbsoluteTime and DiffTime if you do care about leap-seconds.

LeapSecondMap for tracking the leap-seconds

UniversalTime for time based on Earth rotation

Documentation

module Data.Time.Calendar

module Data.Time.Clock

module Data.Time.LocalTime

module Data.Time.Format