Data.Time.Calendar.Easter

Documentation

sundayAfter :: Day -> Day

The next Sunday strictly after a given day.

orthodoxPaschalMoon :: Year -> Day

Given a year, find the Paschal full moon according to Orthodox Christian tradition

orthodoxEaster :: Year -> Day

Given a year, find Easter according to Orthodox Christian tradition

gregorianPaschalMoon :: Year -> Day

Given a year, find the Paschal full moon according to the Gregorian method

gregorianEaster :: Year -> Day

Given a year, find Easter according to the Gregorian method