Data.Time.Calendar.OrdinalDate

Description

ISO 8601 Ordinal Date format

Documentation

data Day

The Modified Julian Day is a standard count of days, with zero being the day 1858-11-17.

type Year = Integer

Year of Common Era (when positive).

type DayOfYear = Int

Day of year, in range 1 (January 1st) to 366. December 31st is 365 in a common year, 366 in a leap year.

type WeekOfYear = Int

Week of year, by various reckonings, generally in range 0-53 depending on reckoning.

pattern YearDay :: Year -> DayOfYear -> Day

Bidirectional abstract constructor for ISO 8601 Ordinal Date format. Invalid day numbers will be clipped to the correct range (1 to 365 or 366).

fromMondayStartWeek

Arguments

:: Year
Year.

-> WeekOfYear
Monday-starting week number (as %W in formatTime).

-> Int
Day of week. Monday is 1, Sunday is 7 (as %u in formatTime).

-> Day

The inverse of mondayStartWeek. Get a Day given the year, the number of the Monday-starting week, and the day of the week. The first Monday is the first day of week 1, any earlier days in the year are week 0 (as %W in formatTime).

fromMondayStartWeekValid

Arguments

:: Year
Year.

-> WeekOfYear
Monday-starting week number (as %W in formatTime).

-> Int
Day of week. Monday is 1, Sunday is 7 (as %u in formatTime).

-> Maybe Day

fromOrdinalDate :: Year -> DayOfYear -> Day

Convert from ISO 8601 Ordinal Date format. Invalid day numbers will be clipped to the correct range (1 to 365 or 366).

fromOrdinalDateValid :: Year -> DayOfYear -> Maybe Day

Convert from ISO 8601 Ordinal Date format. Invalid day numbers return Nothing

fromSundayStartWeek

Arguments

:: Year
Year.

-> WeekOfYear
Sunday-starting week number (as %U in formatTime).

-> Int
Day of week Sunday is 0, Saturday is 6 (as %w in formatTime).

-> Day

The inverse of sundayStartWeek. Get a Day given the year and the number of the day of a Sunday-starting week. The first Sunday is the first day of week 1, any earlier days in the year are week 0 (as %U in formatTime).

fromSundayStartWeekValid

Arguments

:: Year
Year.

-> WeekOfYear
Sunday-starting week number (as %U in formatTime).

-> Int
Day of week. Sunday is 0, Saturday is 6 (as %w in formatTime).

-> Maybe Day

isLeapYear :: Year -> Bool

Is this year a leap year according to the proleptic Gregorian calendar?

mondayStartWeek :: Day -> (WeekOfYear, Int)

Get the number of the Monday-starting week in the year and the day of the week. The first Monday is the first day of week 1, any earlier days in the year are week 0 (as %W in formatTime). Monday is 1, Sunday is 7 (as %u in formatTime).

showOrdinalDate :: Day -> String

Show in ISO 8601 Ordinal Date format (yyyy-ddd)

sundayStartWeek :: Day -> (WeekOfYear, Int)

Get the number of the Sunday-starting week in the year and the day of the week. The first Sunday is the first day of week 1, any earlier days in the year are week 0 (as %U in formatTime). Sunday is 0, Saturday is 6 (as %w in formatTime).

toOrdinalDate :: Day -> (Year, DayOfYear)

Convert to ISO 8601 Ordinal Date format.