Data.Time.Clock

Description

Types and functions for UTC and UT1

Documentation

newtype UniversalTime

The Modified Julian Date is the day with the fraction of the day, measured from UT midnight. It's used to represent UT1, which is time as measured by the earth's rotation, adjusted for various wobbles.

Constructors

ModJulianDate

Fields

getModJulianDate :: Rational

diffTimeToPicoseconds :: DiffTime -> Integer

Get the number of picoseconds in a DiffTime.

picosecondsToDiffTime :: Integer -> DiffTime

Create a DiffTime from a number of picoseconds.

secondsToDiffTime :: Integer -> DiffTime

Create a DiffTime which represents an integral number of seconds.

data DiffTime

This is a length of time, as measured by a clock. Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: DiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)

nominalDay :: NominalDiffTime

One day in NominalDiffTime.

nominalDiffTimeToSeconds :: NominalDiffTime -> Pico

Get the seconds in a NominalDiffTime.

Since: time-1.9.1

secondsToNominalDiffTime :: Pico -> NominalDiffTime

Create a NominalDiffTime from a number of seconds.

Since: time-1.9.1

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

getCurrentTime :: IO UTCTime

Get the current UTCTime from the system clock.

getTime_resolution :: DiffTime

The resolution of getSystemTime, getCurrentTime, getPOSIXTime. On UNIX systems this uses clock_getres, which may be wrong on WSL2.