Data.Time.Calendar.Julian

Year and day format

toJulianYearAndDay :: Day -> (Integer, Int)

Convert to proleptic Julian year and day format. First element of result is year (proleptic Julian calendar), second is the day of the year, with 1 for Jan 1, and 365 (or 366 in leap years) for Dec 31.

fromJulianYearAndDay :: Integer -> Int -> Day

Convert from proleptic Julian year and day format. Invalid day numbers will be clipped to the correct range (1 to 365 or 366).

fromJulianYearAndDayValid :: Integer -> Int -> Maybe Day

Convert from proleptic Julian year and day format. Invalid day numbers will return Nothing

showJulianYearAndDay :: Day -> String

Show in proleptic Julian year and day format (yyyy-ddd)

isJulianLeapYear :: Integer -> Bool

Is this year a leap year according to the proleptic Julian calendar?

toJulian :: Day -> (Integer, Int, Int)

Convert to proleptic Julian calendar. First element of result is year, second month number (1-12), third day (1-31).

fromJulian :: Integer -> Int -> Int -> Day

Convert from proleptic Julian calendar. First argument is year, second month number (1-12), third day (1-31). Invalid values will be clipped to the correct range, month first, then day.

fromJulianValid :: Integer -> Int -> Int -> Maybe Day

Convert from proleptic Julian calendar. First argument is year, second month number (1-12), third day (1-31). Invalid values will return Nothing.

showJulian :: Day -> String

Show in ISO 8601 format (yyyy-mm-dd)

julianMonthLength :: Integer -> Int -> Int

The number of days in a given month according to the proleptic Julian calendar. First argument is year, second is month.

addJulianMonthsClip :: Integer -> Day -> Day

Add months, with days past the last day of the month clipped to the last day. For instance, 2005-01-30 + 1 month = 2005-02-28.

addJulianMonthsRollOver :: Integer -> Day -> Day

Add months, with days past the last day of the month rolling over to the next month. For instance, 2005-01-30 + 1 month = 2005-03-02.

addJulianYearsClip :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th clipped to Feb 28th if necessary. For instance, 2004-02-29 + 2 years = 2006-02-28.

addJulianYearsRollOver :: Integer -> Day -> Day

Add years, matching month and day, with Feb 29th rolled over to Mar 1st if necessary. For instance, 2004-02-29 + 2 years = 2006-03-01.