Data.Time.Calendar.Month

Description

An absolute count of common calendar months.

Documentation

newtype Month

An absolute count of common calendar months. Number is equal to (year * 12) + (monthOfYear - 1).

Constructors

MkMonth Integer

addMonths :: Integer -> Month -> Month

diffMonths :: Month -> Month -> Integer

pattern YearMonth :: Year -> MonthOfYear -> Month

Bidirectional abstract constructor. Invalid months of year will be clipped to the correct range.

fromYearMonthValid :: Year -> MonthOfYear -> Maybe Month

pattern MonthDay :: Month -> DayOfMonth -> Day

Bidirectional abstract constructor. Invalid days of month will be clipped to the correct range.

fromMonthDayValid :: Month -> DayOfMonth -> Maybe Day