Data.Time.Calendar.MonthDay

Documentation

monthAndDayToDayOfYear :: Bool -> Int -> Int -> Int

Convert month and day in the Gregorian or Julian calendars to day of year. First arg is leap year flag.

monthAndDayToDayOfYearValid :: Bool -> Int -> Int -> Maybe Int

Convert month and day in the Gregorian or Julian calendars to day of year. First arg is leap year flag.

dayOfYearToMonthAndDay :: Bool -> Int -> (Int, Int)

Convert day of year in the Gregorian or Julian calendars to month and day. First arg is leap year flag.

monthLength :: Bool -> Int -> Int

The length of a given month in the Gregorian or Julian calendars. First arg is leap year flag.