Data.Time.Calendar.WeekDate

Description

ISO 8601 Week Date format

Documentation

toWeekDate :: Day -> (Integer, Int, Int)

Convert to ISO 8601 Week Date format. First element of result is year, second week number (1-53), third day of week (1 for Monday to 7 for Sunday). Note that "Week" years are not quite the same as Gregorian years, as the first day of the year is always a Monday. The first week of a year is the first week to contain at least four days in the corresponding Gregorian year.

fromWeekDate :: Integer -> Int -> Int -> Day

Convert from ISO 8601 Week Date format. First argument is year, second week number (1-52 or 53), third day of week (1 for Monday to 7 for Sunday). Invalid week and day values will be clipped to the correct range.

fromWeekDateValid :: Integer -> Int -> Int -> Maybe Day

Convert from ISO 8601 Week Date format. First argument is year, second week number (1-52 or 53), third day of week (1 for Monday to 7 for Sunday). Invalid week and day values will return Nothing.

showWeekDate :: Day -> String

Show in ISO 8601 Week Date format as yyyy-Www-d (e.g. "2006-W46-3").