Data.Time.Format.Internal

Description

The contents of this module is liable to change, or disappear entirely. Please let me know if you depend on anything here.

Documentation

class FormatTime t where

Methods

formatCharacter :: Bool -> Char -> Maybe (FormatOptions -> t -> String)

Since: time-1.9.1

class ParseTime t where

The class of types which can be parsed given a UNIX-style time format string.

Minimal complete definition

parseTimeSpecifier, buildTime

Methods

substituteTimeSpecifier :: Proxy t -> TimeLocale -> Char -> Maybe String

Since: time-1.9.1

parseTimeSpecifier :: Proxy t -> TimeLocale -> Maybe ParseNumericPadding -> Char -> ReadP String

Get the string corresponding to the given format specifier.

Since: time-1.9.1

buildTime

Arguments

:: TimeLocale
The time locale.

-> [(Char, String)]
Pairs of format characters and the corresponding part of the input.

-> Maybe t

Builds a time value from a parsed input string. If the input does not include all the information needed to construct a complete value, any missing parts should be taken from 1970-01-01 00:00:00 +0000 (which was a Thursday). In the absence of %C or %Y, century is 1969 - 2068.

Since: time-1.9.1